Update README.md
Siting the Best location for an African Restaurant in Houston, Texas.
Abunuma Gideon
August, 2021

1.	INTRODUCTION
1.1	Background
John and Mary, an immigrant couple from the West African country of Nigeria arrived in the United State last April and are settled in the city of Houston, Texas. Whenever they travel around the city in buses and trains, they notice other passengers of color who were likely of African origin.  Also, at public places, such as parks, clubs and churches, John and Mary encountered people of African descent. Their observation of such high number of peoples from their common African culture got them thinking that they could use this peculiar human resource to further their economic interest and start a business in Houston similar to the one they had back home while living in Lagos, Nigeria: An African Restaurant. 
1.2	Problem
Among the many challenges John and Mary expect to face while setting up their business is the decision to select a location to site their proposed restaurant location. To help resolve this problem they have contacted their friend, Peter, who is a data scientist to help the use location data to find the main areas of Houston Texas where people from African origin live and work. These people will represent the bulk of their customers and patrons for their proposed restaurant. Peter will attempt to use foursquare location data to identify the demographics of the Houston area and to find the appropriate location of the proposed John and Mary restaurant to serve the people of African origins in the Houston area of Texas.
1.3	Interest
Besides John and Mary, other persons such as tourists and business patrons are also interested in using location data to find business places relevant to their passion and interests such as the African restaurant which John and Mary desire to build.
2.	Data Acquisition and Cleaning
The demographic data of the Houston area of Texas can be found from the U.S Census 2020 data which is replicated in the foursquare application.  The dataset shows the ethnic representations of people living in Houston such as Hispanics, Asian-Americans, whites and people of African descent in terms of their geographical spread and concentration. The purpose of the cleaning process will be to isolate the Houston areas which have a decent majority of people of African descent. This data processing and analysis will help to effectively locate the proposed African restaurant
3.	Exploratory Data Analysis
Further data analysis of the data involves the disaggregation of the data to identify possible patrons of African cuisine. This is done by the conduction of survey in the geographical area among respondents and residents to get an idea of the number of persons that will be open to buying African foods.
4.	Conclusion
In this study, we analysed demographic and location data to identify the best possible locations to site an African restaurant. The final decision is arrived at based on the concurrent proportional relationship in the number of people of African descent living in proximity to the proposed African restaurant and the population of other respondents and residents who are willing to try out, and for African Foods.